backend/requirements.txt
Flask==3.1.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
PyJWT==2.9.0
python-dotenv==1.1.1
eventlet==0.40.1